[#6] Move homepage including tests to locale route dir
